close #68, close #56

+++ src/main/java/com/adventurealley/adventurexp/controllers/ActivityController.java
@@ -2,12 +2,15 @@ package com.adventurealley.adventurexp.controllers;
 
 import com.adventurealley.adventurexp.dtos.CreateActivityDTO;
 import com.adventurealley.adventurexp.entities.Activity;
+import com.adventurealley.adventurexp.entities.Role;
+import com.adventurealley.adventurexp.entities.User;
 import com.adventurealley.adventurexp.services.IActivityService;
 import org.json.JSONArray;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @RestController()
@@ -20,9 +23,17 @@ public class ActivityController
     }
 
     @PostMapping("/api/activities")
-    ResponseEntity<String> createActivity(@RequestBody CreateActivityDTO createActivityDTO) {
+    ResponseEntity<String> createActivity(@RequestBody CreateActivityDTO createActivityDTO, HttpSession httpSession) {
+        User u = (User)httpSession.getAttribute("user");
+
+        if (u == null) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
+        if (u.getRole() != Role.EMPLOYEE) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
         Activity a = activityService.createActivity(createActivityDTO);
         return new ResponseEntity<>(a.toJSON().toString(), HttpStatus.CREATED);
+
+
     }
 
     @GetMapping("/api/activities")
@@ -44,7 +55,13 @@ public class ActivityController
     }
 
     @DeleteMapping("/api/activities/{id}")
-    ResponseEntity<String> deleteById (@PathVariable("id") long id) {
+    ResponseEntity<String> deleteById (@PathVariable("id") long id, HttpSession httpSession) {
+        User u = (User)httpSession.getAttribute("user");
+
+        if (u == null) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
+        if (u.getRole() != Role.EMPLOYEE) return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
         activityService.delete( id );
         return null;
